# api/openapi.yaml
openapi: 3.1.0
info:
  title: sing-box-panel API
  version: 0.1.0
  license:
    name: GPL-3.0-or-later
    identifier: GPL-3.0-or-later
servers:
  - url: /api/v1
paths:
...
  /auth/session:
    get:
      operationId: getSession
      responses:
        "200":
          description: Current browser or bearer session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Problem"
    post:
      operationId: createSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: A new browser session. The management token is never returned.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteSession
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      responses:
        "204":
          description: The current browser session was removed.
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
  /system/status:
    get:
      operationId: getSystemStatus
      responses:
        "200":
          description: Machine-readable runtime and control-plane status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStatus"
        "401":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /dashboard/context:
    get:
      operationId: getDashboardContext
      responses:
        "200":
          description: Human-facing summary for the initial panel dashboard.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardContext"
        "401":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
...
  /config/canonical:
    get:
      operationId: getCanonicalConfiguration
      responses:
        "200":
          description: Current immutable canonical snapshot.
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanonicalSnapshot"
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
    put:
      operationId: replaceCanonicalConfiguration
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - name: If-Match
          in: header
          required: true
          description: Quoted base revision ID, or `"none"` for the first revision.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CanonicalDocument"
      responses:
        "200":
          description: Canonical head after the compare-and-swap save.
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanonicalSave"
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "428": { $ref: "#/components/responses/Problem" }
    patch:
      operationId: patchCanonicalConfiguration
      description: Applies one ordered set of lossless JSON-pointer changes in a single canonical compare-and-swap revision.
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - name: If-Match
          in: header
          required: true
          description: Quoted base revision ID.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CanonicalPatchRequest" }
      responses:
        "200":
          description: Canonical head after applying every change atomically.
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CanonicalSave" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "428": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    managementBearer:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: sbp_session
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required for cookie-authenticated state-changing requests.
      schema:
        type: string
        minLength: 1
...
  responses:
    Problem:
      description: Stable RFC 9457-compatible problem response.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 1
          writeOnly: true
    Session:
      type: object
      additionalProperties: false
      required: [displayName]
      properties:
        displayName:
          type: string
        csrfToken:
          type: string
          description: Present for cookie sessions so the browser can make writes.
        expiresAt:
          type: string
          format: date-time
    SystemStatus:
      type: object
      additionalProperties: false
      required:
        - panel_version
        - canonical_revision
        - running
        - capability_state
      properties:
        panel_version:
          type: string
        canonical_revision:
          type: integer
          format: int64
          minimum: 0
        applied_bundle_id:
          type: [string, "null"]
        running:
          type: boolean
        running_version:
          type: [string, "null"]
        running_artifact:
          type: [string, "null"]
        capability_state:
          type: string
          enum: [unavailable, unresolved, native_structured, compatible_structured, manual_json, quarantined_manual_json]
          description: Exact-version capability state for the running core, or the applied bundle while stopped; never an assumed latest-version state.
    DashboardContext:
      type: object
      additionalProperties: false
      required: [view, running, canonical, applied, capability]
      properties:
        view:
          type: object
          additionalProperties: false
          required: [exactVersion]
          properties:
            exactVersion:
              type: string
        running:
          oneOf:
            - type: "null"
            - type: object
              additionalProperties: false
              required: [exactVersion, artifactName, digest]
              properties:
                exactVersion: { type: string }
                artifactName: { type: string }
                digest: { type: string }
        canonical:
          type: object
          additionalProperties: false
          required: [revision, savedAt, hasUnappliedChanges]
          properties:
            revision: { type: integer, format: int64, minimum: 0 }
            savedAt: { type: string, format: date-time }
            hasUnappliedChanges: { type: boolean }
        applied:
          oneOf:
            - type: "null"
            - type: object
              additionalProperties: false
              required: [bundle, revision, appliedAt]
              properties:
                bundle: { type: string, minLength: 1 }
                revision: { type: integer, format: int64, minimum: 1 }
                appliedAt: { type: string, format: date-time }
        capability:
          type: object
          additionalProperties: false
          required: [level, label, warning]
          properties:
            level:
              type: string
              enum: [native_structured, compatible_structured, manual_json, unavailable]
            label: { type: string }
            warning: { type: [string, "null"] }
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        request_id: { type: string }
        fields:
          type: object
          additionalProperties: { type: string }
    CanonicalDocument:
      type: object
      additionalProperties: false
      required: [schema_version, global, nodes, rules, subscription]
      properties:
        schema_version: { type: integer, const: 1 }
        global: { type: object, additionalProperties: true }
        nodes:
          type: array
          maxItems: 10000
          items:
            type: object
            required: [id, kind, enabled]
            properties:
              id: { type: string }
              kind: { type: string }
              enabled: { type: boolean }
            additionalProperties: true
        rules:
          type: array
          maxItems: 10000
          items:
            type: object
            required: [id, enabled]
            properties:
              id: { type: string }
              enabled: { type: boolean }
            additionalProperties: true
        subscription: { type: object, additionalProperties: true }
    CanonicalSnapshot:
      type: object
      additionalProperties: false
      required: [id, sequence, schema_version, document, document_json, sha256, created_at]
      properties:
        id: { type: string }
        sequence: { type: integer, format: int64, minimum: 1 }
        parent_id: { type: string }
        schema_version: { type: integer, const: 1 }
        document: { $ref: "#/components/schemas/CanonicalDocument" }
        document_json:
          type: string
          description: Exact canonical JSON bytes encoded as a JSON string for lossless browser parsing.
        sha256: { type: string, pattern: "^[0-9a-f]{64}$" }
        created_at: { type: string, format: date-time }
    CanonicalChange:
      type: object
      additionalProperties: false
      required: [op, path]
      properties:
        op:
          type: string
          enum: [set, unset]
        path: { type: string, minLength: 1, maxLength: 4096 }
        value_json:
          type: string
          description: Required for set and forbidden for unset; contains exactly one JSON value without numeric coercion.
    CanonicalPatchRequest:
      type: object
      additionalProperties: false
      required: [changes]
      properties:
        changes:
          type: array
          minItems: 1
          maxItems: 4096
          items: { $ref: "#/components/schemas/CanonicalChange" }
    CanonicalSave:
      type: object
      additionalProperties: false
      required: [revision, no_change]
      properties:
        revision: { $ref: "#/components/schemas/CanonicalSnapshot" }
        task_id: { type: string }
        no_change: { type: boolean }
...
security:
  - sessionCookie: []
  - managementBearer: []